pcb/pyproject.toml
[tool.poetry]
name = "pcb"
version = "0.1.0"
description = ""
authors = ["Martin Karlsch <martin@karlsch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kigadgets = "^0.4.2"
kicad-skip = "^0.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
